Get CM properties only when required
Currently, initialization of drm device would fail if there is no color
management properties (CTM, gamma_lut and gamma_lut_size), which is the
case for virtio-GPU.  But actually these properties are not required when
color adjustment is disabled.

Only try to get these properties when color adjustment is required.

Tracked-On: OAM-118492
Signed-off-by: Weifeng Liu <weifeng.liu@intel.com>
Signed-off-by: Li, HaihongX <haihongx.li@intel.com>